lollms_core/requirements.txt
tqdm
pyyaml
Pillow
wget
setuptools
requests


safe_store[all]
pipmaster
ascii_colors>=0.1.3
beautifulsoup4

autopep8

fastapi
uvicorn
python-multipart
python-socketio
python-socketio[client]
python-socketio[asyncio_client]

freedom-search
scrapemaster